requirements.txt
Django==3.2.4
pip==21.1.1
wheel==0.36.2
pytz==2021.1
sqlparse==0.4.1
asgiref==3.3.4
setuptools==56.0.0
djangorestframework==3.12.4